Do actually fix the original issue #1208, fix the "all Delta Chat screen stays white" problem
In order to actually fix the original issue, I moved the screen locking from onStart() to onResume(). The problem had been that applyScreenLock uses startActivityForResult() to show the screen lock. Therefore, isWaitingForResult was set to true and shouldLock() returned false when called from onPause(). As onActivityResult() sets isWaitingForResult to false, and onResume() is called after onActivityResult(), this problem does not exist anymore. (now applyScreenLock is called twice in this case, from onActivityResult() and from onResume(), but this should be fine because as soon as one screen lock gets the right credentials, onActivityResult() is called again with RESULT_OK and the second screen lock is never shown to the user).

I also fixed the issue of the previous commit: (once the lock is disabled again in the system (via Delta Chat/Settings/App access/Change secret), all Delta Chat screen stays white unrecoverable, also after restarts.)

The problem had been that shouldLock() does call isScreenLockAvailable() in the end but isScreenLockAvailable() only checked the system version. Now it also checks keyguardManager.isKeyguardSecure() i.e. if there is a keyguard that could be called. To unify things, AppProtectionPreferenceFragment also calls this method now instead of checking this same thing itself.